# Use std::variant instead of mapbox::util::variant