--- a/include/engine/api/base_result.hpp
+++ b/include/engine/api/base_result.hpp
@@ -2,7 +2,7 @@
 #define ENGINE_API_BASE_RESULT_HPP
 
 #include <flatbuffers/flatbuffers.h>
-#include <mapbox/variant.hpp>
+#include <variant>
 
 #include <string>
 
@@ -11,7 +11,7 @@
 namespace osrm::engine::api
 {
 using ResultT =
-    mapbox::util::variant<util::json::Object, std::string, flatbuffers::FlatBufferBuilder>;
+    std::variant<util::json::Object, std::string, flatbuffers::FlatBufferBuilder>;
 } // namespace osrm::engine::api
 
 #endif