--- a/include/engine/api/table_api.hpp
+++ b/include/engine/api/table_api.hpp
@@ -420,8 +420,8 @@ class TableAPI final : public BaseAPI
                       [&](const auto &cell)
                       {
                           util::json::Array row;
-                          row.values.emplace_back(util::json::Number(cell.row));
-                          row.values.emplace_back(util::json::Number(cell.column));
+                          row.values.emplace_back(util::json::Value{util::json::Number(cell.row)});
+                          row.values.emplace_back(util::json::Value{util::json::Number(cell.column)});
                           json_table.values.emplace_back(util::json::Value{std::move(row)});
                       });
         return json_table;