--- a/include/engine/api/table_api.hpp
+++ b/include/engine/api/table_api.hpp
@@ -377,7 +377,9 @@ class TableAPI final : public BaseAPI
                                return util::json::Value(
                                    util::json::Number(from_alias<double>(duration) / 10.));
                            });
-            json_table.values.push_back(util::json::Value{std::move(json_row)});
+
+            util::json::Value tmp{std::move(json_row)};
+            json_table.values.push_back(std::move(tmp));
         }
         return json_table;
     }