--- a/include/engine/api/table_api.hpp
+++ b/include/engine/api/table_api.hpp
@@ -418,11 +418,11 @@ class TableAPI final : public BaseAPI
         util::json::Array json_table;
         std::for_each(fallback_speed_cells.begin(),
                       fallback_speed_cells.end(),
-                      [&](const auto &)
+                      [&](const auto &cell)
                       {
                           util::json::Array row;
-                        //   row.values.push_back(
-                        //       util::json::Value{util::json::Number(static_cast<double>(cell.row))});
+                          util::json::Value jCellRow{util::json::Number(static_cast<double>(cell.row))};
+                          row.values.push_back(jCellRow);
                         //   row.values.push_back(util::json::Value{
                         //       util::json::Number(static_cast<double>(cell.column))});
                           json_table.values.push_back(util::json::Value{row});