--- a/include/extractor/maneuver_override.hpp
+++ b/include/extractor/maneuver_override.hpp
@@ -11,7 +11,7 @@
 #include "util/std_hash.hpp"
 #include "util/vector_view.hpp"
 
-#include <mapbox/variant.hpp>
+#include <variant>
 
 #include <algorithm>