--- a/include/extractor/restriction.hpp
+++ b/include/extractor/restriction.hpp
@@ -1,10 +1,10 @@
 #ifndef RESTRICTION_HPP
 #define RESTRICTION_HPP
 
+#include "turn_path.hpp"
 #include "util/coordinate.hpp"
 #include "util/opening_hours.hpp"
 #include "util/typedefs.hpp"
-#include "turn_path.hpp"
 #include <limits>
 
 namespace osrm::extractor