--- a/include/server/service/base_service.hpp
+++ b/include/server/service/base_service.hpp
@@ -5,7 +5,7 @@
 #include "osrm/osrm.hpp"
 #include "util/coordinate.hpp"
 
-#include <mapbox/variant.hpp>
+#include <variant>
 
 #include <string>
 #include <vector>