--- a/include/util/geojson_debug_logger.hpp
+++ b/include/util/geojson_debug_logger.hpp
@@ -51,7 +51,8 @@ class GeojsonLogger
             if (!first)
                 ofs << ",\n\t\t";
 
-            util::json::render(ofs, object.get<util::json::Object>());
+            (void)object;
+           // util::json::render(ofs, std::get<util::json::Object>(object));
 
             first = false;
         }