We get rid of old unsupported dependency this way + even gain some performance according to some of benchmarks.

<!-- BENCHMARK_RESULTS_START -->
## Benchmark Results
| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1156.49<br/>plain u32: 1165.4<br/>aliased double: 1221.16<br/>plain double: 1194.4 | aliased u32: 1173.03<br/>plain u32: 1169.35<br/>aliased double: 1205.46<br/>plain double: 1205.09 |
| json-render | String: 6.82666ms<br/>Stringstream: 10.6403ms<br/>Vector: 6.72893ms | String: 6.54718ms<br/>Stringstream: 9.33254ms<br/>Vector: 6.87933ms |
| match_ch | Default radius: <br/>4.52153ms/req at 82 coordinate<br/>0.0551406ms/coordinate<br/>Radius 5m: <br/>4.51387ms/req at 82 coordinate<br/>0.0550472ms/coordinate<br/>Radius 10m: <br/>15.311ms/req at 82 coordinate<br/>0.186719ms/coordinate<br/>Radius 15m: <br/>37.246ms/req at 82 coordinate<br/>0.454219ms/coordinate<br/>Radius 30m: <br/>318.143ms/req at 82 coordinate<br/>3.8798ms/coordinate | Default radius: <br/>4.60197ms/req at 82 coordinate<br/>0.0561215ms/coordinate<br/>Radius 5m: <br/>4.81285ms/req at 82 coordinate<br/>0.0586933ms/coordinate<br/>Radius 10m: <br/>15.2201ms/req at 82 coordinate<br/>0.185611ms/coordinate<br/>Radius 15m: <br/>37.3425ms/req at 82 coordinate<br/>0.455396ms/coordinate<br/>Radius 30m: <br/>317.949ms/req at 82 coordinate<br/>3.87743ms/coordinate |
| match_mld | Default radius: <br/>2.8728ms/req at 82 coordinate<br/>0.0350341ms/coordinate<br/>Radius 5m: <br/>2.88596ms/req at 82 coordinate<br/>0.0351947ms/coordinate<br/>Radius 10m: <br/>10.6612ms/req at 82 coordinate<br/>0.130015ms/coordinate<br/>Radius 15m: <br/>26.8552ms/req at 82 coordinate<br/>0.327502ms/coordinate<br/>Radius 30m: <br/>314.604ms/req at 82 coordinate<br/>3.83663ms/coordinate | Default radius: <br/>2.82204ms/req at 82 coordinate<br/>0.0344151ms/coordinate<br/>Radius 5m: <br/>2.81336ms/req at 82 coordinate<br/>0.0343093ms/coordinate<br/>Radius 10m: <br/>10.4019ms/req at 82 coordinate<br/>0.126853ms/coordinate<br/>Radius 15m: <br/>26.7546ms/req at 82 coordinate<br/>0.326275ms/coordinate<br/>Radius 30m: <br/>312.855ms/req at 82 coordinate<br/>3.8153ms/coordinate |
| packedvector | random write:<br/>std::vector 11404.9 ms<br/>util::packed_vector 74442.1 ms<br/>slowdown: 6.5272<br/>random read:<br/>std::vector 11162.7 ms<br/>util::packed_vector 31109.9 ms<br/>slowdown: 2.78697 | random write:<br/>std::vector 11509.3 ms<br/>util::packed_vector 78345 ms<br/>slowdown: 6.8071<br/>random read:<br/>std::vector 11427.6 ms<br/>util::packed_vector 34020 ms<br/>slowdown: 2.97701 |
| route_ch | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>595.822ms<br/>0.595822ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>381.316ms<br/>0.381316ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>755.935ms<br/>0.755935ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>153.334ms<br/>0.153334ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>98.6903ms<br/>0.0986903ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>134.79ms<br/>0.13479ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>153.34ms<br/>0.15334ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>98.5865ms<br/>0.0985865ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>134.563ms<br/>0.134563ms/req | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>515.086ms<br/>0.515086ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>355.196ms<br/>0.355196ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>635.942ms<br/>0.635942ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>152.472ms<br/>0.152472ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>97.3469ms<br/>0.0973469ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>132.701ms<br/>0.132701ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>152.579ms<br/>0.152579ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>97.9879ms<br/>0.0979879ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>132.23ms<br/>0.13223ms/req |
| route_mld | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>728.132ms<br/>0.728132ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>467.816ms<br/>0.467816ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>947.211ms<br/>0.947211ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>284.93ms<br/>0.28493ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>165.884ms<br/>0.165884ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>302.766ms<br/>0.302766ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>281.587ms<br/>0.281587ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>166.466ms<br/>0.166466ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>303.502ms<br/>0.303502ms/req | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>647.257ms<br/>0.647257ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>441.606ms<br/>0.441606ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>816.873ms<br/>0.816873ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>276.712ms<br/>0.276712ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>163.881ms<br/>0.163881ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>298.151ms<br/>0.298151ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>275.659ms<br/>0.275659ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>163.1ms<br/>0.1631ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>298.177ms<br/>0.298177ms/req |
| rtree | 1 result:<br/>209.484ms ->  0.0209484 ms/query<br/>10 results:<br/>244.768ms ->  0.0244768 ms/query | 1 result:<br/>208.05ms ->  0.020805 ms/query<br/>10 results:<br/>242.531ms ->  0.0242531 ms/query |
<!-- BENCHMARK_RESULTS_END -->