@jcoupey 2024-11-07
I guess I'm a bit late to the party here, but this merge results in a breaking change to the C++ API, see [adjustments to the example file](https://github.com/Project-OSRM/osrm-backend/pull/6903/files#diff-6b7a0e0510d76cfcff77f45d4ef0b7b4ab758394340b4f50bdd4403a476da67c).

I definitely see the benefit of dropping a custom implementation to use something supported in the standard, it's just that historically the OSRM project has been very cautious in avoiding breaking changes unless bumping the major release number. Does this mean the next release would have to be v6 something, or simply that the non-breaking policy is somewhat loosened?