@SiarheiFedartsou 2024-11-07
@jcoupey I think worth trying to target v6 - I believe it is not the only breaking change we made lately. There were too many problems with mapbox::variant(e.g. C++20 compatibility, performance)... But it in general are there that many libosrm users out there who would rely on us to be super backward compatible? 🤔 
Do you use libosrm directly in VROOM btw?