@jcoupey 2024-11-07
> are there that many libosrm users out there

Not sure, but since you never really hear complaints until you break semver, you never know. ;-)

I do use libosrm in VROOM which is why I commented. Downstream we would have to make the change to support the next release, but that would mean dropping support for previous OSRM versions which is not ideal for users. Or we would have to implement some ugly workaround to adjust the calls based on libosrm version, not ideal either.

Either way since OSRM adheres to semver (and especially if there are other breaking changes), it would make sense to bump the major number...